Split out and connect MiniCartHeader.
